// api/database/constants.ts
export enum TableName {
  Permissions = 'permissions',
  Users = 'users',
}

export enum UserField {
  Id = 'id',
  Email = 'email',
  PasswordHash = 'password_hash',
  Username = 'username',
  FirstName = 'first_name',
  LastName = 'last_name',
  CreatedAt = 'created_at',
  UpdatedAt = 'updated_at',
  DeletedAt = 'deleted_at',
}